Update js and set image as block
